+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Additional labels for pre-release and build metadata are available as extensions
 is not equal to the derived temporal resolution.
 - Added `data_resolution` argument to `average_data_by_period`, `monthly_means`, `coverage` and 
   `merge_datasets_by_period` functions (Issue #[297](https://github.com/brightwind-dev/brightwind/issues/297))
+- Update to work with Pandas 1.3.2. This mostly includes depreciating pd.Timedelta and using pd.DateOffset instead. (Pull request [#312](https://github.com/brightwind-dev/brightwind/pull/312)).
 
 ## [2.0.0]
 - Major changes, notably